requirements.txt
RPI.GPIO==0.6.3
